old-pullr/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
discord==1.7.3
discord.py==1.7.3
Flask==2.1.2
idna==3.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
Pillow==9.2.0
python-dotenv==0.20.0
requests==2.28.1
soupsieve==2.3.2.post1
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
yarl==1.7.2
